backend/auth: move API endpoints into their own file
This is for better organization (endpoint code might grow in the future).
