+++ src/backend/expungeservice/endpoints/auth.py
@@ -5,10 +5,9 @@ import re
 
 from flask.views import MethodView
 from flask import request, abort, jsonify, current_app
-from werkzeug.security import check_password_hash, generate_password_hash
+from werkzeug.security import check_password_hash
 
-# TODO temp hack - replace with table from DB
-users = {}
+from expungeservice.endpoints.users import users
 
 def get_auth_token(app, username):
     dt = datetime.datetime.utcnow()
@@ -56,26 +55,6 @@ def auth_required(f):
             return 'auth token expired!', 401
     return wrapper
 
-class Users(MethodView):
-    def post(self):
-        data = request.get_json()
-
-        if data == None:
-            abort(400)
-
-        # TODO temp hack - replace with table from DB
-        if data['username'] in users:
-            return 'User already exists! Please login', 202
-
-        users[data['username']] = generate_password_hash(data['password'])
-
-        response_data = {
-            'username': data['username'],
-            'email address': data['email address']
-        }
-
-        return jsonify(response_data), 201
-
 class AuthToken(MethodView):
     def get(self):
         data = request.get_json()
@@ -92,12 +71,5 @@ class AuthToken(MethodView):
         }
         return jsonify(response_data)
 
-class ProtectedView(MethodView):
-    @auth_required
-    def get(self):
-        return 'Protected View'
-
 def register(app):
-    app.add_url_rule('/api/v0.1/users', view_func=Users.as_view('users'))
     app.add_url_rule('/api/v0.1/auth_token', view_func=AuthToken.as_view('auth_token'))
-    app.add_url_rule('/api/v0.1/test/protected', view_func=ProtectedView.as_view('protected'))
